Crawlers/SourceCodeCrawler/requirements.txt
py-etherscan-api
natsort